add separate backend adapter for ease of testing
